move content to JSON file
