+++ src/components/Teaser/index.tsx
@@ -1,54 +1,5 @@
-import { useEffect, useRef, useState, type ReactElement } from 'react'
-import { ButtonBase } from '@mui/material'
-import css from './styles.module.css'
-import { ALPHA_TELEGRAM_LINK } from '@/config/constants'
+import type { ReactElement } from 'react'
+import PageContent from '@/components/common/PageContent'
+import teaserContent from '@/content/teaser.json'
 
-export const Teaser = (): ReactElement => {
-  const [ready, setReady] = useState(false)
-  const [showButton, setShowButton] = useState(false)
-
-  const videoRef = useRef<HTMLVideoElement>(null)
-
-  const handleVideoEnd = () => {
-    setShowButton(true)
-  }
-
-  useEffect(() => {
-    const poll = setInterval(() => {
-      if (!videoRef.current) return
-      if (videoRef.current.readyState >= videoRef.current.HAVE_FUTURE_DATA) {
-        setReady(true)
-        clearInterval(poll)
-      }
-    }, 100)
-
-    return () => clearInterval(poll)
-  }, [])
-
-  return (
-    <div className={css.container}>
-      {!showButton ? (
-        <div className={css.wrapper}>
-          <video
-            autoPlay
-            muted
-            playsInline
-            ref={videoRef}
-            style={{ opacity: 0 }}
-            onEnded={handleVideoEnd}
-            className={`${css.video} ${ready ? css.ready : ''}`}
-          >
-            {/* TODO: Replace by Teaser_with_logo.mp4 */}
-            <source src="/videos/Teaser/Teaser.mp4" type="video/mp4" />
-          </video>
-        </div>
-      ) : (
-        <div className={css.imageWrapper}>
-          <ButtonBase target="_blank" rel="noreferrer" href={ALPHA_TELEGRAM_LINK} className={css.button}>
-            <img src="/images/Teaser/telegram.svg" alt="Telegram" />
-          </ButtonBase>
-        </div>
-      )}
-    </div>
-  )
-}
+export const Teaser = (): ReactElement => <PageContent content={teaserContent} path="teaser.json" />
