+++ src/content/teaser.json
@@ -1,0 +1,10 @@
+[
+  {
+    "pageTitle": "Ethereum Smart Accounts",
+    "description": "Own the Internet. Ethereum Smart Accounts to safeguard your digital assets and build the future of ownership.",
+    "component": "common/MetaTags"
+  },
+  {
+    "component": "Teaser/Video"
+  }
+]
